Converts project to Gradle and adds a .gitignore
