+++ build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'java'
+}
+
+group "me.improperissues"
+version "1.2.4"
+
+repositories {
+    maven { url = uri("https://repo.papermc.io/repository/maven-public/") }
+}
+
+dependencies {
+    compileOnly("io.papermc.paper:paper-api:1.17.1-R0.1-SNAPSHOT")
+}
