--- main/resources/config.yml
@@ -1,26 +1,0 @@
-config:
-  plugin:
-    prefix: '§7[§3§lUni§b§lV§7] §'
-    watermark: 'Plugin by ImproperIssues, visit "github.com/ItziSpyder/UniVault"'
-  player:
-    max_speed: 10.0
-    command_delay: 1
-    delay_blacklist:
-      - /review
-      - /submit
-      - /vault
-      - /archive
-  general:
-    join_clear: false
-    max_pages: 50
-    max_nbt_length: 5000
-    min_nbt_length: 10
-  permissions:
-    requires_op: false
-    nonop_edit: true
-  submissions:
-    enabled: true
-    duplicates: false
-    cooldown: 10
-  archive:
-    can_break_or_place: false
